completely reworked and cleaned backend endpoints
Split endpoints to GET and POST to individual files accordingly
moved commonly used functions to new helper script, imports on appropriate places
added some comments
some other minor housekeeping
