+++ server.js
@@ -3,19 +3,19 @@ require("dotenv").config();
 const app = express();
 const cors = require("cors");
 const mysql = require("mysql2");
-const jwt = require("jsonwebtoken");
 
-//API Gets
-//app.use(express.json())
+// custom logger
 function log(req, res, next) {
     console.log("Got a request type: " + req.method);
     console.warn("Got: " + req.originalUrl);
     //console.log(req.headers)
     next();
 }
-  
+
+// get port on which service should be available from env
 const port = process.env.PORT;
 
+// create database connection
 const connection = mysql.createConnection({
     host: process.env.DB_HOST,
     user: process.env.DB_USER,
@@ -23,9 +23,11 @@ const connection = mysql.createConnection({
     password: process.env.DB_PASS
 });
 
+// use custom logger and set up CORS
 app.use(log);
 app.use(cors({ origin: ["https://opwngrid.xyz","https://api.opwngrid.xyz"] }));
 
+// TODO: ideally completely drop this
 app.get("*", function(req, res, next){ 
     if(req.headers.host == "opwngrid.xyz") { //if it's a sub-domain
         if (req.url.includes("/api/")) {
@@ -48,13 +50,15 @@ app.get("*", function(req, res, next){
 });
 
 // all routes leading to static content here
-require("./routes/static_routes.js")(app);
+// disabled, since frontend moved to its own repo
+//require("./routes/static_routes.js")(app);
 
 // all routes regarding statistics
 require("./routes/statistics_routes.js")(app, connection);
 
 // all routes regarding API
-require("./routes/api_routes.js")(app, connection);
+require("./routes/api_GET_routes.js")(app, connection);
+require("./routes/api_POST_routes.js")(app, connection);
 
 process.on("SIGINT", function() {
     connection.end(function(err) {
